activitymap.py: "To:" in a commit is not a real interaction
Don't count "To:" in the body of a commit as actually doing anything as
it is just added by developers to notify people, they don't do any work
on their end.

It might also end up sucking in mailing list aliases, which you don't
want to track anywhere as they are not "real".