activitymap.py: signed-off-by: does NOT mean co-authored-by
When a commit is signed-off, it does NOT mean that the person was the
author, they merely signed-off on it, that is all.  So fix up the
mappings to properly show this.

The authorship information should be coming directly from the author of
the git id, and maybe a few other fields in the signed-off-by area like
co-developed and such.